% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monet-metadata.R
\name{AddGSEMData}
\alias{AddGSEMData}
\title{Add GSEM Data to Monet Object}
\usage{
AddGSEMData(monet_object, id.gse, id.gsm)
}
\arguments{
\item{monet_object}{A Monet object.}

\item{id.gse}{The GSE ID.}

\item{id.gsm}{A vector of GSM IDs to associate with the GSE.}
}
\value{
The updated Monet object.
}
\description{
Add GSM IDs under a GSE in the Monet object.
}
