% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monet-metadata.R
\name{AddGSMData}
\alias{AddGSMData}
\title{Add GSM Data to Monet Object}
\usage{
AddGSMData(monet_object, gsm_data)
}
\arguments{
\item{monet_object}{A Monet object.}

\item{gsm_data}{A data frame containing GSM metadata to add.}
}
\value{
The updated Monet object.
}
\description{
Add or update GSM metadata in the Monet object.
}
