% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_df.R
\name{anova_df}
\alias{anova_df}
\title{Wrapper for anova() returning a single data.frame}
\usage{
anova_df(..., colnames = "auto")
}
\arguments{
\item{...}{passed to \code{\link{anova}}}

\item{colnames}{Determines which columns from \code{as.data.frame(anova(...))}
will be collected. \code{"auto"} attempts to figure out which columns are
relevant via heuristics. Alternatively, a list of character strings. The
first element gives the name(s) of the \code{df} columns,
the second element the column containing the statistic,
the third element, th ecolumn containing the p-value.}
}
\value{
\code{data.frame} with one row and at least three columns.
}
\description{
Wrapper for anova() returning a single data.frame
}
