% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_coancestries.R
\name{calculate_coancestries}
\alias{calculate_coancestries}
\title{calculate coancestries}
\usage{
coancestries_calculate(genetic_data_parents, genetic_data_F1, population_individual_id = NA, pedigree = NA, data_type = "dosage", BiAllelic = TRUE)
}
\arguments{
\item{genetic_data_parents}{A data frame or file name containing genetic data for the parental generation. Must be dosage format.}

\item{genotyped_parent_populations}{A vector of population IDs for the parental generation.}

\item{genetic_data_F1}{A data frame or file name containing genetic data for the F1 generation. Must be dosage format.}

\item{population_individual_id}{(Optional) A data frame mapping individuals to their respective populations. First parent ids, then F1 ids.
If not provided, a balanced design is assumed.}

\item{pedigree}{(Optional) A data frame containing pedigree information with at least three columns:
individual ID, dam ID, and sire ID.}

\item{all_parents_genotyped}{Logical; if TRUE, indicates that the genotyped individuals in "genotyped_parent_populations" are the parents of F1 individuals}
}
\value{
A list containing:
\item{The.M}{A kinship-based relatedness matrix adjusted for F1 individuals.}
\item{Theta.P}{An adjusted population coancestry matrix.}
}
\description{
Given genetic data it provides coancestry for between and within populations
}
\details{
This function calculates kinship and coancestry measures from genetic data
or pedigree-based information, adjusting for population structure and F1 generation characteristics.
}
\references{
\itemize{
\item Goudet & Weir (2023)
}
}
\author{
Isabela do O \email{isabela.doo@unil.ch}
}
