% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{checkFile}
\alias{checkFile}
\title{Check if file exists}
\usage{
checkFile(file_path)
}
\arguments{
\item{file_path}{character. String specifying location of file to be read.}
}
\description{
Check if the specified file exists and give better error message.
}
