% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting_blocks_matrix.r
\name{counting_blocks_matrix}
\alias{counting_blocks_matrix}
\title{Counting blocks of non-zero rows in a matrix}
\usage{
counting_blocks_matrix(mat)
}
\arguments{
\item{mat}{A matrix where rows are considered for counting blocks}
}
\value{
data frame of the block number, the number of rows in the block, and the number of columns in the block
}
\description{
This function counts the number of blocks of non-zero rows in a given matrix.
A block is defined as a contiguous sequence of non-zero rows.
}
