% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monet-class.R
\name{createMonetObject}
\alias{createMonetObject}
\title{Monet Object Constructor}
\usage{
createMonetObject(
  assays = list(),
  meta.data = list(),
  active.assay = NULL,
  misc = list()
)
}
\arguments{
\item{assays}{A list of assays (expression data layers).}

\item{meta.data}{A list of metadata layers (gse.index, gsm.index, pmc.index).}

\item{active.assay}{The active assay for operations.}

\item{misc}{Additional information.}
}
\value{
A Monet object.
}
\description{
Create a Monet object for GEO data management.
}
