% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data for LAVA package}
\format{
A data frame with placeholder structure:
\describe{
\item{sample_id}{Sample identifier}
\item{population}{Population identifier}
\item{variance}{Variance measurement}
\item{ancestral_group}{Ancestral group classification}
}
}
\source{
This is placeholder documentation for future data
}
\usage{
example_data
}
\description{
A dataset containing example genetic variance data for demonstration purposes.
This is a placeholder for future data that will be included in the package.
}
\keyword{datasets}
