% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{filePath}
\alias{filePath}
\title{Wrapper around file.exists}
\usage{
filePath(file_path)
}
\arguments{
\item{file_path}{File path to be tested.}
}
\value{
file_path or "NA"
}
\description{
Need NA return for monet if file does not exist.
}
