% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monet-filter.R
\name{filter_stdd}
\alias{filter_stdd}
\title{Filter and standardise gene expression}
\usage{
filter_stdd(monet_data, no_genes = NULL, noise_flt = NULL,
  mu_th = 0.1, sd_th = 0.1)
}
\arguments{
\item{monet_data}{monetData class.}

\item{no_genes}{integer. No of genes to filter out based on high signal to
noise in gene expression.}

\item{noise_flt}{numeric. Threshold for signal to noise cutoff.}

\item{mu_th}{numeric. Default \code{=0.1}. Remove genes with average gene
expression across time \code{< mu_th}}

\item{sd_th}{numeric. Default \code{=0.1}. Remove genes with standard deviation
of gene expression across time.}
}
\description{
Filter and standardise gene expression
}
