% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoNET.R
\name{Gene_Meta_query}
\alias{Gene_Meta_query}
\title{Query shortest paths from genes to metabolites}
\usage{
Gene_Meta_query(
  MulOmicNet,
  genelist,
  metalist,
  genefilter = "",
  metafilter = "",
  minpath = 1,
  maxpath = 3
)
}
\arguments{
\item{MulOmicNet}{A dataframe of network with two columns.}

\item{genelist}{A dataframe of genes with one column.}

\item{metalist}{A dataframe of metabolites with one column.}

\item{genefilter}{Character constant,the type of input gene ID.Gene ID includes Gene_Entrez_ID, Gene_Name, Ensembl_Gene_ID, Enzyme_IDs and RefSeq_IDs.
For the detail,please see the function ListFilters.}

\item{metafilter}{Character constant, the type of input metabolite ID.Metabolite ID includes Metabolites, Names, Formulas, HMDB_ID, KEGG_ID, PubChem_ID and CHEBI_ID.
For the detail,please see the function ListFilters.}

\item{minpath}{The lower limit of the shortest path length,the default value is 1.}

\item{maxpath}{The upper limit of the shortest path length,the default value is 3.}
}
\value{
A list is returned, each list element contains a shortest path from one gene to one metabolite.
Also a text file is generated with the list written to.
}
\description{
Taking genes and metabolites as input, one can get paths in our multi-omic network from genes to metabolites.
}
\examples{
MulOmicNet<-BuildNet(threshold=0.5)
genelist<- as.data.frame(c("A1BG","AOC3"))
metalist<- as.data.frame(c("bamppald[c]","Water"))
PathList <- Gene_Meta_query(MulOmicNet=MulOmicNet,genelist=genelist,metalist=metalist,
                            genefilter="Gene_Name",metafilter="Names",minpath=1,maxpath=3)

}
