% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{geneColCheck}
\alias{geneColCheck}
\title{geneColCheck}
\usage{
geneColCheck(data_dt)
}
\arguments{
\item{data_dt}{data.table. Data in \code{data.table} format}
}
\description{
if no gene col passed check if character column exists otherwise use index
}
