% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoNET.R
\name{Genequery}
\alias{Genequery}
\title{Query shortest paths from/to genes}
\usage{
Genequery(
  MulOmicNet,
  genelist,
  filter = "",
  node = c("metabolite", "snp"),
  minpath = 1,
  maxpath = 3
)
}
\arguments{
\item{MulOmicNet}{A dataframe of network with two columns}

\item{genelist}{A dataframe of genes with one column.}

\item{filter}{Character constant,the type of input gene ID.Gene ID includes Gene_Entrez_ID, Gene_Name, Ensembl_Gene_ID, Enzyme_IDs and RefSeq_IDs.
For the detail,please see the function ListFilters.}

\item{node}{Character constant, gives the shortest paths between genes to which type of nodes should be calculated.
If 'metabolite',the default,then the shortest paths from genes to metabolites will be calculated.
If 'snp',then the shortest paths from SNPs to genes will be calculated.}

\item{minpath}{the lower limit of the shortest path length,the default value is 1.}

\item{maxpath}{the upper limit of the shortest path length,the default value is 3.}
}
\value{
A list is returned, each list element contains a shortest path from one SNP to one gene or from one gene to one metabolite.
Also a text file is generated with the list written to.
}
\description{
Taking genes as input, one can get paths in our multi-omic network from SNPs to genes or from genes to metabolites.
}
\examples{
MulOmicNet<-BuildNet(threshold=0.5)
genelist<- as.data.frame(c("A1BG","A1BG-AS1","A2M","A2ML1"))
PathList<-Genequery(MulOmicNet=MulOmicNet,genelist=genelist,filter="Gene_Name",
                    node="metabolite",minpath=1,maxpath=3)

}
