% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{geneStanderdise}
\alias{geneStanderdise}
\title{standardise gene expression}
\usage{
geneStanderdise(gene_long)
}
\arguments{
\item{gene_long}{data.table.}
}
\description{
standardise gene expression in long format.
}
