% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{getAtacSeq,monetData-method}
\alias{getAtacSeq,monetData-method}
\title{Extract ATACseq slot}
\usage{
\S4method{getAtacSeq}{monetData}(.Object)
}
\arguments{
\item{monetData}{monetData class object.}
}
\description{
Extract ATACseq slot
}
