% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{getAtacSeq,monetDataRaw-method}
\alias{getAtacSeq,monetDataRaw-method}
\title{Extract ATACseq slot}
\usage{
\S4method{getAtacSeq}{monetDataRaw}(.Object)
}
\arguments{
\item{monetDataRaw}{monetDataRaw class object.}
}
\description{
Extract ATACseq slot
}
