% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{getGeneExp,monetData-method}
\alias{getGeneExp,monetData-method}
\title{Extract gene expression slot}
\usage{
\S4method{getGeneExp}{monetData}(.Object)
}
\arguments{
\item{monetData}{monetData class object.}
}
\description{
Extract gene expression slot
}
