% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{getGeneExp,monetDataRaw-method}
\alias{getGeneExp,monetDataRaw-method}
\title{Extract gene expression slot}
\usage{
\S4method{getGeneExp}{monetDataRaw}(.Object)
}
\arguments{
\item{monetDataRaw}{monetDataRaw class object.}
}
\description{
Extract gene expression slot
}
