% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{getGeneInit,monetDataRaw-method}
\alias{getGeneInit,monetDataRaw-method}
\title{Extract initial gene expression}
\usage{
\S4method{getGeneInit}{monetDataRaw}(.Object)
}
\arguments{
\item{monetDataRaw}{monetDataRaw class object.}
}
\description{
Extract gene expression for the initial time-point.
}
