% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monet-metadata.R
\name{GetGSEMetadata}
\alias{GetGSEMetadata}
\title{Get GSE Metadata}
\usage{
GetGSEMetadata(monet_object)
}
\arguments{
\item{monet_object}{A Monet object.}
}
\value{
A data frame containing GSE metadata.
}
\description{
Retrieve GSE metadata from the Monet object.
}
