% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monet-metadata.R
\name{GetPMCMetadata}
\alias{GetPMCMetadata}
\title{Get PMC Metadata}
\usage{
GetPMCMetadata(monet_object, pmc_id = NULL)
}
\arguments{
\item{monet_object}{A Monet object.}

\item{pmc_id}{Optional PMC ID for specific query. If NULL, return all PMC metadata.}
}
\value{
A data frame containing PMC metadata.
}
\description{
Retrieve PMC metadata from the Monet object.
}
