% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{initialize,monetData-method}
\alias{initialize,monetData-method}
\title{initialize monetData}
\usage{
\S4method{initialize}{monetData}(.Object, monetDataRaw, gene_filt)
}
\description{
initialize monetData
}
