% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{initialize,monetDataFilt-method}
\alias{initialize,monetDataFilt-method}
\title{initialize monetDataFilt}
\usage{
\S4method{initialize}{monetDataFilt}(.Object, monetData, gene_filt)
}
\description{
initialize monetDataFilt
}
