% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{initialize,monetDataRaw-method}
\alias{initialize,monetDataRaw-method}
\title{initilize monetDataRaw}
\usage{
\S4method{initialize}{monetDataRaw}(.Object, gene_dt, atac_dt, gene_nm,
  gene_t0_dt, gene_path = "NA", atac_path = "NA")
}
\arguments{
\item{monetDataRaw}{}
}
\value{
monetDataRaw object
}
\description{
initilize monetDataRaw
}
