% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{initialize,monetOptim-method}
\alias{initialize,monetOptim-method}
\title{initialize monetOptim class}
\usage{
\S4method{initialize}{monetOptim}(.Object, optim_out, monet_dat)
}
\arguments{
\item{monetOptim}{}
}
\value{
monetOptim object
}
\description{
initialize monetOptim class
}
