% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_numeric_vector}
\alias{is_numeric_vector}
\title{Check if input is a numeric vector}
\usage{
is_numeric_vector(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical value indicating if x is a numeric vector
}
\description{
This function checks if the provided input is a numeric vector.
}
\examples{
\dontrun{
is_numeric_vector(1:10)  # TRUE
is_numeric_vector("test")  # FALSE
}
}
\keyword{internal}
