% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_positive_numeric}
\alias{is_positive_numeric}
\title{Check if input is a positive numeric vector}
\usage{
is_positive_numeric(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical value indicating if x is a positive numeric vector
}
\description{
This function checks if the provided input is a numeric vector with all positive values.
}
\examples{
\dontrun{
is_positive_numeric(1:10)  # TRUE
is_positive_numeric(c(-1, 2, 3))  # FALSE
}
}
\keyword{internal}
