% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lava.r
\name{lava}
\alias{lava}
\title{lava: Log ratios of ancestral variances}
\usage{
lava(Theta.P, The.M, trait_dataframe, column_individual = "id", column_trait = "trait", ...)
}
\arguments{
\item{Theta.P}{A square matrix representing the coancestry matrix between populations.}

\item{The.M}{A square matrix representing the kinship-based relatedness matrix for individuals.}

\item{trait_dataframe}{A data frame containing individual IDs and trait values.
The first column should be individual IDs, and the second column should be trait values.}

\item{column_individual}{The name of the column containing individual IDs. Default is "id".}

\item{column_trait}{The name of the column containing trait values. Default is "trait".}

\item{...}{Additional arguments passed to the brms function.}
}
\value{
A lava type object containing:
\item{posterior_samples}{A list with posterior samples of variance components and residuals.}
\item{BRMS_stats}{A list with the median, lower and upper bounds of the 95\% credible interval for the variance difference.}
\item{log_ratio}{A list with the p-value of the log ratios, the mean of the log ratio of between- and within-population variance, and confidence intervals.}
\item{BRMS_model}{A list with the fitted Bayesian model and hypothesis test results.}
}
\description{
Given coancestry matrices for between and within populations and a trait data frame,
this function estimates the log ratio of ancestral variances using a Bayesian mixed-effects model.
}
\details{
This function standardizes trait data, constructs a Bayesian mixed-effects model
using \code{brms}, and estimates ancestral variances. The function assumes no crosses between populations
and analyzes one trait at a time.
}
\references{
\itemize{
\item Goudet & Weir (2023)
\item do O et al (2025)
}
}
\author{
Isabela do O \email{isabela.doo@unil.ch}
}
