% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoNET.R
\name{ListFilters}
\alias{ListFilters}
\title{Query the types of gene/metabolite ID}
\usage{
ListFilters(what)
}
\arguments{
\item{what}{Character constant, 'gene'and 'metabolite' represent the querying types of gene ID and metabolite ID respectively.}
}
\value{
character vector
}
\description{
Provide users to look up available input ID for genes or metabolites.
}
\details{
Gene ID includes Gene_Entrez_ID, Gene_Name, Ensembl_Gene_ID, Enzyme_IDs and RefSeq_IDs.
Gene_Entrez_ID: stable identifiers for genes at the NCBI (e.g.,51166).
Gene_Name: the official gene symbol that has been approved by the HGNC (e.g.,AADAT).
Ensembl_Gene_ID:curated ensembl gene ID (e.g.,ENSG00000109576).
Enzyme_IDs:Enzyme entries have Enzyme Commission (EC) numbers associated with them that indicate the hierarchical
functional classes to which they belong. This field can contain multiple values as a comma delimited list (e.g.,2.6.1.39, 2.6.1.7).
RefSeq_IDs:The Reference Sequence (RefSeq) identifier for that entry, provided by the NCBI. This field may contain multiple
values as a comma delimited list (e.g.,NM_016228).
Metabolite ID includes Metabolites, Names, Formulas, HMDB_ID, KEGG_ID, PubChem_ID and CHEBI_ID.
Metabolites:descriptive name of metabolite (e.g.,h2o[c]).
Names:full names of metabolite (e.g.,Water).
Formulas:fuomula of metabolite (e.g.,H2O).
HMDB_ID:identifier of metaboite from human metabolome database (e.g.,HMDB02111).
KEGG_ID:identifier metabolite from KEGG database (e.g.,C00001).
PubChem_ID:identifier of metabolite from PubChem database (e.g.,962).
CHEBI_ID:identifier of metabolite from Chemical Entities of Biological Interest(ChEBI)database (e.g.,15377).
}
\examples{
ListFilters("gene")

}
