% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoNET.R
\name{Metaquery}
\alias{Metaquery}
\title{Query shortest paths to metabolites}
\usage{
Metaquery(
  MulOmicNet,
  metalist,
  filter = "",
  node = c("gene", "snp"),
  minpath = 1,
  maxpath = 3
)
}
\arguments{
\item{MulOmicNet}{A dataframe of network with two columns.}

\item{metalist}{A dataframe of metabolites with one column.}

\item{filter}{Character constant, the type of input metabolite ID.Metabolite ID includes Metabolites, Names, Formulas, HMDB_ID, KEGG_ID, PubChem_ID and CHEBI_ID.
For the detail,please see the function ListFilters.}

\item{node}{character constant, gives the shortest paths from which type of nodes to metabolites should be calculated.
If 'gene',the default,then the shortest paths from genes to metabolites will be calculated.
If 'snp',then the shortest paths from SNPs to metabolites will be calculated.}

\item{minpath}{the lower limit of the shortest path length,the default value is 1.}

\item{maxpath}{the upper limit of the shortest path length,the default value is 3.}
}
\value{
A list is returned, each list element contains a shortest path from one SNP to one metabolite or from one gene to one metabolite.
Also a text file is generated with the list written to.
}
\description{
Taking metabolites as input, one can get paths in our multi-omic network from SNPs or from genes to metabolites.
}
\examples{
MulOmicNet<-BuildNet(threshold=0.5)
metalist<- as.data.frame(c("10-Formyltetrahydrofolate","Water"))
PathList<-Metaquery(MulOmicNet=MulOmicNet,metalist=metalist,filter="Names",
                    node="gene",minpath=1,maxpath=2)

}
