% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{monet-class}
\alias{monet-class}
\title{monetDataSuper superclass}
\description{
monetDataSuper superclass
}
\section{Slots}{

\describe{
\item{\code{path}}{to gene expression}

\item{\code{path}}{to ATACseq data}
}}

