% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monet-package.R
\docType{package}
\name{monet}
\alias{monet}
\alias{monet-package}
\title{The 'monet' package.}
\description{
\strong{M}ulti-\strong{O}mics \strong{NET}work inference (MONET)
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
