% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monet-input.R
\name{monet_input_prep}
\alias{monet_input_prep}
\title{Monet input prep}
\usage{
monet_input_prep(monet_class, par_list = list())
}
\arguments{
\item{monet_class}{monet data class. Possibilities include \code{monetData} or
\code{monetDataFilt} classes.}

\item{par_list}{list. This list obhject should contain any predefined
parameters, this needs to be a names list. The following named parameters
can be passed: \code{no_tf}, \code{de_b}, \code{sigma}, \code{sigma_x}, \code{alpha}}
}
\value{
list for monet stan model.
}
\description{
Prepare input monet class for Rstan model.
}
