% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monet-run.R
\name{monet_run}
\alias{monet_run}
\title{Run monet}
\usage{
monet_run(gene_exp = NULL, atac_seq = NULL, ..., gene_col = NULL,
  gene_atac = NULL, gene_exp0 = NULL, gene_filt = NULL,
  noise_flt = NULL, no_genes = NULL, par_list = list(),
  trnsfrm_ge = c("asinh", "log", "none"))
}
\arguments{
\item{gene_exp}{}
}
\description{
This is the main
}
