% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{monetData-class}
\alias{monetData-class}
\title{MONET filtered data input class}
\description{
MONET filtered data input class
}
\section{Slots}{

\describe{
\item{\code{exp_filt}}{filtered list of gene}
}}

