% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monet-data.R
\name{monetData_init}
\alias{monetData_init}
\title{monetData initialize}
\usage{
monetData_init(gene_exp = NULL, atac_seq = NULL, gene_col = NULL,
  transform = c("asinh", "log", "none"), gene_atac = NULL,
  gene_exp0 = NULL)
}
\arguments{
\item{gene_exp}{gene expression. Either path to gene expression file or
data.frame of gene expression. Rows represent genes and columns represent}

\item{atac_seq}{ATAC-Seq.}

\item{gene_col}{column of gene names in gene_exp either in the file or
data.frame}

\item{gene_atac}{column of gene names in atac_seq if different from
gene_col.}
}
\value{
monetData class based on input to function.
}
\description{
Function to initialise monetData class
}
\examples{

}
