% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{monetOptim-class}
\alias{monetOptim-class}
\title{MONET optimization output class}
\description{
MONET optimization output class
}
\section{Slots}{

\describe{
\item{\code{x_fit}}{matrix}

\item{\code{b_fit}}{matrix}

\item{\code{w}}{vector.}
}}

