% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_model_formulas.R
\name{nested_model_formulas}
\alias{nested_model_formulas}
\title{Create nested models formulas and corresponding data.frame}
\usage{
nested_model_formulas(formula, data, extra_formula, type = 3,
  test_intercept = FALSE, na.action)
}
\arguments{
\item{formula}{Two-sided formula. Separate formulas (i.e., nested models) will be created for each model term on right side of this formula. For example, for mixed-effects models this should be the response (i.e., dependent) variable plus the fixed-effects.}

\item{data}{\code{data.frame} passed to \code{model.frame} holding the variables in \code{formula} and \code{extra_formula}.}

\item{extra_formula}{Optional one-sided formula that will be added to the resulting formulas. No separate formulas will be built based on this formula. For example, for mixed-effects model this should contain the random-effects.}

\item{type}{Type of sums of squares. Currently only Type III (i.e., \code{3} or \code{"III"}) are supported.}

\item{test_intercept}{Logical. Should test for intercept be included in output? Default is \code{FALSE}}

\item{na.action}{how \code{NA}s are treated. Passed to \code{\link{model.frame}}.}
}
\value{
List with three slots:
\describe{
  \item{\code{formulas}}{List of named formulas. First entry is the \code{full_model}, each following entry contains the formula for the full model after withholding one model term.}
  \item{\code{data}}{\code{data.frame} holding all variables used inthe  \code{formulas}}
  \item{\code{reduced_formulas}}{Same as formulars, but without adding the \code{extra_formula}.}
}
}
\description{
Create nested models formulas and corresponding data.frame
}
\author{
Henrik Singmann with contributions from \href{http://stackoverflow.com/q/11335923/289572}{Ben Bolker and Joshua Wiley}.
}
