% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neutral_data_g3000}
\alias{neutral_data_g3000}
\title{Neutral Genetic Data (Parental Generation)}
\format{
A data file in FSTAT format containing:
\describe{
\item{Populations}{18 populations (10 individuals each)}
\item{Loci}{2000 neutral bi-allelic markers}
\item{Format}{FSTAT format: population ID followed by genotypes}
}
}
\source{
Simulated neutral marker data
}
\usage{
system.file("extdata", "neutral_data_g3000.dat", package = "MONET")
}
\description{
Neutral bi-allelic genetic marker data for 180 individuals from 18 populations
(10 individuals per population). This dataset represents the parental generation
used to create F1 offspring for common garden experiments.
}
\details{
This file contains neutral genetic markers simulated to represent population
structure without selection. The data is in FSTAT format and can be read
using hierfstat::read.fstat().

After reading, the data should be converted to dosage format using
hierfstat::biall2dos() for use with calculate_coancestries().

The file structure:
\itemize{
\item First line: n_populations n_loci n_digits_per_allele max_alleles_per_locus
\item Subsequent lines: population_id genotype_locus1 genotype_locus2 ...
}
}
\examples{
\dontrun{
# Read the data
neutral_file <- system.file("extdata", "neutral_data_g3000.dat", 
                            package = "MONET")
sim <- hierfstat::read.fstat(fname = neutral_file)

# Convert to dosage format
dos <- hierfstat::biall2dos(sim[, -1])
pop <- sim$Pop

# Use in coancestry calculation
# See vignette for complete workflow
}

}
\keyword{datasets}
