% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice.R
\name{nice}
\alias{nice}
\alias{nice.monet}
\alias{print.nice_table}
\title{Make nice ANOVA table for printing.}
\usage{
nice(object, ...)

\method{nice}{monet}(object, sig_symbols = attr(object$anova_table,
  "sig_symbols"), ...)

\method{print}{nice_table}(x, ...)
}
\arguments{
\item{object, x}{An object of class \code{"monet"}.}

\item{...}{currently ignored.}

\item{sig_symbols}{Character. What should be the symbols designating
significance? When entering an vector with \code{length(sig.symbol) < 4}
only those elements of the default (\code{c(" +", " *", " **", " ***")})
will be replaced. \code{sig_symbols = ""} will display the stars but not
the \code{+}, \code{sig_symbols = rep("", 4)} will display no symbols.
The default is given by \code{afex_options("sig_symbols")}.}
}
\description{
This generic function produces a nice ANOVA table for objects of class
 \code{"monet"} as returned by \code{\link{test_terms}()}.
}
\note{
Same functionality as the function with the same name from \pkg{afex}.
}
