% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monet-parse.R
\name{ParseGSEMetadata}
\alias{ParseGSEMetadata}
\title{Parse GSE Metadata}
\usage{
ParseGSEMetadata(file_path)
}
\arguments{
\item{file_path}{The path to the GSE soft file.}
}
\value{
A data frame containing the parsed GSE metadata.
}
\description{
Parse metadata from a GSE soft file.
}
