% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monet-parse.R
\name{ParseGSESoft}
\alias{ParseGSESoft}
\title{Parse GSE Metadata}
\usage{
ParseGSESoft(file_path)
}
\arguments{
\item{file_path}{The path to the GSE soft file.}
}
\value{
A data frame containing the parsed GSE metadata.
}
\description{
Parse soft file from a GSE soft file.
}
