% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monet-parse.R
\name{ParseGSMMeta}
\alias{ParseGSMMeta}
\title{Parse GSM Metadata}
\usage{
ParseGSMMeta(file_path)
}
\arguments{
\item{file_path}{The path to the GSM soft file.}
}
\value{
A data frame containing the parsed GSM metadata.
}
\description{
Parse metadata from a GSM soft file.
}
