% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMR.R
\name{PlotMR}
\alias{PlotMR}
\title{Barplot of Mitochondrial and/or Ribosomal Genes in a Seurat Gene Set}
\usage{
PlotMR(obj, feature)
}
\arguments{
\item{obj}{A gene expression matrix or Seurat assay matrix with gene names as rownames.}

\item{feature}{A character string specifying which genes to plot.
Must be one of \code{"mito"}, \code{"ribo"}, or \code{"both"}.}
}
\value{
A \code{ggplot2} barplot showing the number of mitochondrial and/or ribosomal genes.
}
\description{
This function identifies mitochondrial and/or ribosomal genes within a gene expression matrix
or Seurat assay matrix and visualizes their counts using a barplot.
It can be used to assess the presence of these gene categories before filtering.
Must be used before ReMove, or number of features found will be 0.
}
\examples{
\dontrun{
  PlotMR(seurat_object@assays$RNA@counts, feature = "mito")
  PlotMR(seurat_object, feature = "both")
}

}
