% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPCA.R
\name{PlotPCA}
\alias{PlotPCA}
\title{Plot Explained Variance of Principal Components from a Seurat object}
\usage{
PlotPCA(obj, dim = 10)
}
\arguments{
\item{obj}{A Seurat object that contains a PCA reduction (created with \code{RunPCA}).}

\item{dim}{An integer indicating how many top principal components to plot.}
}
\value{
A \code{ggplot2} object showing a barplot of explained variance per component.
}
\description{
This function plots the proportion of variance explained by the top \code{dim} principal components
from a Seurat object's PCA reduction.
}
\details{
The function computes the variance explained by each principal component using the standard deviation
values stored in \code{obj@reductions$pca@stdev}, then plots the first \code{dim} components.
Default \code{dim} is 10.
}
\examples{
\dontrun{
  PlotPCA(seurat_object, dim = 10)
}

}
