% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monet-data.R
\name{prepAtac}
\alias{prepAtac}
\title{prepAtac}
\usage{
prepAtac(atac_seq, gene_col = NULL)
}
\arguments{
\item{atac_seq}{file path or data.}

\item{gene_col}{character (defaul is \code{NULL}). Column name (file or data
object) containing gene names.}
}
\description{
prepare ATAC seq data for usage in monet
}
