% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monet-data.R
\name{prepGeneExp}
\alias{prepGeneExp}
\title{prepGeneExp}
\usage{
prepGeneExp(gene_exp = NULL, gene_col = NULL)
}
\arguments{
\item{gene_exp}{data or path to data.file.}

\item{gene_col}{gene name column.}
}
\description{
Prepare gene expression data for usage in monet
}
