% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quanti_trait_g3000}
\alias{quanti_trait_g3000}
\title{Quantitative Trait Loci Data (Parental Generation)}
\format{
A data file in FSTAT format containing:
\describe{
\item{Populations}{18 populations (10 individuals each)}
\item{Loci}{100 bi-allelic QTL markers}
\item{Format}{FSTAT format: population ID followed by genotypes}
}
}
\source{
Simulated QTL data with additive genetic effects
}
\description{
Bi-allelic genetic marker data for quantitative trait loci (QTL) from 180
individuals across 18 populations. This dataset represents causal loci
underlying a quantitative trait in the parental generation.
}
\details{
This file contains genetic markers at loci that influence a quantitative trait.
Unlike the neutral markers in neutral_data_g3000.dat, these loci have phenotypic
effects and are used to simulate trait values in the F1 generation.

The data is in FSTAT format and should be read using hierfstat::read.fstat(),
then converted to dosage format with hierfstat::biall2dos().

These markers can be used to:
\itemize{
\item Simulate trait values in offspring
\item Understand the genetic architecture of traits
\item Create realistic scenarios for testing local adaptation
}
}
\examples{
\dontrun{
# Read the data
quanti_file <- system.file("extdata", "quanti_trait_g3000.dat", 
                           package = "MONET")
sim_quanti <- hierfstat::read.fstat(fname = quanti_file)

# Convert to dosage format
dos_quanti <- hierfstat::biall2dos(sim_quanti[, -1], diploid = TRUE)
}

}
\keyword{datasets}
