% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{set_key_dt}
\alias{set_key_dt}
\title{Set key in data.table}
\usage{
set_key_dt(data_dt, col_name)
}
\arguments{
\item{data_dt}{data.table. A \code{data.table} object.}

\item{col_name}{the column containing the key variable.}
}
\description{
helper function for \code{\link[data.table:setkeyv]{setkeyv()}} function for more
useful error messages in the context of monet.
}
