% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{setDataTest,monetData-method}
\alias{setDataTest,monetData-method}
\title{Update data_test slot}
\usage{
\S4method{setDataTest}{monetData}(.Object, data_test)
}
\arguments{
\item{data_test}{new data_test entry.}

\item{monetData}{monetData class object.}
}
\description{
Update data_test slot
}
