% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{setGeneExp,monetDataRaw-method}
\alias{setGeneExp,monetDataRaw-method}
\title{Update gene expression}
\usage{
\S4method{setGeneExp}{monetDataRaw}(.Object, gene_dt)
}
\arguments{
\item{gene_dt}{data.table. New entry to be used. First column with gene
names (named "gene") subsequent columns are gene expression at
time-points.}

\item{monetDataRaw}{S4 class.}
}
\description{
Update gene expression
}
