% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{setGeneInit,monetDataRaw-method}
\alias{setGeneInit,monetDataRaw-method}
\title{Update init gene expression}
\usage{
\S4method{setGeneInit}{monetDataRaw}(.Object, gene_init)
}
\arguments{
\item{gene_init}{data.table. New entry to be used. First column with gene
names (named "gene") second column with the entries.}

\item{monetDataRaw}{S4 class.}
}
\description{
Update init gene expression
}
