% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{show,monetData-method}
\alias{show,monetData-method}
\title{Print method for monetData}
\usage{
\S4method{show}{monetData}(object)
}
\arguments{
\item{monetData}{}
}
\description{
Print method for monetData
}
