% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoNET.R
\name{SNP_Meta_query}
\alias{SNP_Meta_query}
\title{Query shortest paths from SNPs to Metabolites}
\usage{
SNP_Meta_query(
  MulOmicNet,
  snplist,
  metalist,
  filter = "",
  minpath = 1,
  maxpath = 3
)
}
\arguments{
\item{MulOmicNet}{A dataframe of network with two columns.}

\item{snplist}{A dataframe of SNPs with one column.}

\item{metalist}{a dataframe of metabolites with one column.}

\item{filter}{Character constant, the type of input metabolite ID.Metabolite ID includes Metabolites, Names, Formulas, HMDB_ID, KEGG_ID, PubChem_ID and CHEBI_ID.
For the detail,please see the function ListFilters.}

\item{minpath}{the lower limit of the shortest path length,the default value is 1.}

\item{maxpath}{the upper limit of the shortest path length,the default value is 3.}
}
\value{
A list is returned, each list element contains a shortest path from one SNP to one metabolite.
Also a text file is generated with the list written to.
}
\description{
Taking SNPs and metabolites as input, one can get paths in our multi-omic network from SNPs to metabolites.
}
\examples{
MulOmicNet<-BuildNet(threshold=0.5)
snplist<- as.data.frame(c("rs1000313","rs1001104"))
metalist<- as.data.frame(c("10-Formyltetrahydrofolate","Water"))
PathList<-SNP_Meta_query(MulOmicNet=MulOmicNet,snplist=snplist,metalist=metalist,
                         filter="Names",minpath=6,maxpath=7)

}
