% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoNET.R
\name{SNPquery}
\alias{SNPquery}
\title{Query shortest paths from SNPs}
\usage{
SNPquery(
  MulOmicNet,
  snplist,
  node = c("gene", "metabolite"),
  minpath = 1,
  maxpath = 3
)
}
\arguments{
\item{MulOmicNet}{A dataframe of network with two columns.}

\item{snplist}{A dataframe of SNP IDs with one column.}

\item{node}{Character constant, gives the shortest paths from SNP to which type of nodes should be calculated.
If 'gene',the default,then the shortest paths from SNPs to genes will be calculated.
If 'metabolite',then the shortest paths from SNPs to metabolites will be calculated.}

\item{minpath}{the lower limit of the shortest path length,the default value is 1.}

\item{maxpath}{the upper limit of the shortest path length,the default value is 3.}
}
\value{
A list is returned, each list element contains a shortest path from one SNP to one gene or from one SNP to one metabolite.
Also a text file is generated with the list written to.
}
\description{
Taking SNPs as input, one can get paths in our multi-omic network from SNPs to genes or metabolites.
}
\examples{
MulOmicNet<-BuildNet(threshold=0.5)
snplist<- as.data.frame(c("rs1000313","rs1001106"))
PathList<-SNPquery(MulOmicNet=MulOmicNet,snplist=snplist,node="gene",minpath=3,maxpath=4)

}
