% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoNET.R
\name{SubNetRFW}
\alias{SubNetRFW}
\title{Print output}
\usage{
SubNetRFW(MulOmicNet, PathList, Anno = FALSE, enrich = FALSE, simplot = FALSE)
}
\arguments{
\item{MulOmicNet}{A dataframe of network with two columns.}

\item{PathList}{A list of paths that generated by query function.}

\item{Anno}{Logical scalar,whether to annotate the nodes in the identified paths. the default is FALSE.}

\item{enrich}{Logical scalar,whether to perform the gene enrichment analysis, the default is FALSE.}

\item{simplot}{Logical scalar, whether to plot the subnetwork, the default is FALSE.}
}
\value{
text files and PDF file
}
\description{
Print results to text files, plot subnetwork and perform gene enrichiment analysis.
}
\details{
A subnetwork composed of nodes involved in the identified paths is generated. The subnetwork
is written into a text file with each line representing one interaction.When argument Anno is set to true,
one can get annotation files of nodes involved in the identified paths.When argument enrich is set to true,
all genes involved in the identified paths are enriched in the 2019 KEGG human pathway database provided by Enrichr.
When argument simplot is set to true, an image of the subnetwork in PDF format is generated. The subnetwork is plotted
by placing vertices on the plane using the force-directed layout algorithm.
}
\examples{
MulOmicNet<-BuildNet(threshold=0.5)
snplist<- as.data.frame(c("rs1000313","rs1001104"))
metalist<- as.data.frame(c("10-Formyltetrahydrofolate","Water"))
PathList<-SNP_Meta_query(MulOmicNet=MulOmicNet,snplist=snplist,metalist=metalist,
                         filter="Names",minpath=6,maxpath=7)
SubNetRFW(MulOmicNet=MulOmicNet,PathList=PathList,Anno=TRUE,enrich=TRUE,simplot=TRUE)

}
