% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermSelect.R
\name{TermSelect}
\alias{TermSelect}
\title{Select Genes by Biotype from a GTF Annotation matching a Seurat object expression or Expression Matrix}
\usage{
TermSelect(obj, gtf, type)
}
\arguments{
\item{obj}{A gene expression matrix with gene names as rownames or Seurat object.}

\item{gtf}{A data frame representing gene annotations (e.g., parsed GTF),
containing gene names and gene biotype columns.}

\item{type}{A character string specifying the gene biotype to select (e.g., "protein_coding").}
}
\value{
A data frame or Seurat object containing gene names and biotypes filtered by the specified biotype also contained in the Seurat object or matrix.
}
\description{
This function filters gene annotations from a GTF-like data frame to select
only genes of a specified biotype present in a given Seurat object or gene expression matrix.
}
\examples{
\dontrun{
  pc_genes <- TermSelect(seurat_object@assays$RNA@counts, gtf_annotation_df, "protein_coding")
  lnc_genes <- TermSelect(seurat_object, gtf_annotation_df, "LNC")
}

}
