% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{testGeneCol}
\alias{testGeneCol}
\title{testing gene_col var}
\usage{
testGeneCol(gene_col, file_path)
}
\arguments{
\item{gene_col}{character or integer. The column which contains gene names.}

\item{file_path}{character. Location of file to be read.}
}
\description{
Testing if the gene_col variable provided is contained within the data.
Also converts numerical column indicator into characters.
}
