% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fun.R
\name{trnsfrmGe}
\alias{trnsfrmGe}
\title{Transforming data function}
\usage{
trnsfrmGe(data_dt, trnsfrm_ge)
}
\arguments{
\item{data_dt}{data.table. A \code{data.table} object of gene expression.}

\item{trnsfrm_ge}{character.}
}
\description{
helper function
}
