% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monet-metadata.R
\name{AddGSEMData}
\alias{AddGSEMData}
\title{Add GSEM Data to Monet Object (Append)}
\usage{
AddGSEMData(monet_object, gse_id, gsm_ids)
}
\arguments{
\item{monet_object}{A Monet object.}

\item{gse_id}{The GSE ID to be added.}

\item{gsm_ids}{A vector of GSM IDs to be associated with the GSE.}
}
\value{
Updated Monet object with the appended GSE-GSM mapping.
}
\description{
Add GSE and GSM mapping to the Monet object, appending to existing data.
}
