% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\docType{methods}
\name{initialize,monetData-method}
\alias{initialize,monetData-method}
\title{initilize monetData}
\usage{
\S4method{initialize}{monetData}(.Object, gene_dt, atac_dt, gene_nm,
  gene_t0_dt, gene_path = "NA", atac_path = "NA")
}
\arguments{
\item{monetData}{}
}
\value{
monetData object
}
\description{
initilize monetData
}
